perf(build)!: slice portable contract capabilities
Split Runtime Ports and Tool Contracts into owner-scoped additive features, update every workspace consumer, and keep owner-specific tests isolated. Add fail-closed Cargo and source guards for target identity, feature graphs, dependency edges, forwarding, activation, and hidden aggregates.

BREAKING CHANGE: direct consumers of bitfun-runtime-ports or bitfun-agent-tools must explicitly enable the capability features they use.
