+++ scripts/core-boundaries/self-test.mjs
@@ -507,6 +507,8 @@ export function runManifestParserSelfTest({
         'dep:anyhow',
         'dep:async-trait',
         'dep:bitfun-runtime-ports',
+        'bitfun-runtime-ports/runtime-event-port',
+        'bitfun-runtime-ports/workspace-ports',
         'dep:dunce',
         'process-runtime',
         'tokio/fs',
@@ -599,6 +601,7 @@ export function runManifestParserSelfTest({
     [
       'client',
       [
+        'bitfun-agent-tools/acp-bridge',
         'dep:futures',
         'dep:serde',
         'dep:bitfun-core',
@@ -609,6 +612,7 @@ export function runManifestParserSelfTest({
     [
       'server',
       [
+        'dep:bitfun-agent-tools',
         'dep:bitfun-agent-runtime',
         'dep:bitfun-core-types',
         'dep:bitfun-core',
@@ -2465,7 +2469,7 @@ export function runManifestParserSelfTest({
 
   const requiredContentContracts = [
     {
-      path: 'src/crates/contracts/runtime-ports/src/lib.rs',
+      path: 'src/crates/contracts/runtime-ports/src/agent_api.rs',
       contracts: [
         'AgentDialogTurnRequest',
         'AgentDialogPrependedReminder',
