+++ src/apps/cli/Cargo.toml
@@ -55,10 +55,10 @@ bitfun-core-types = { path = "../../crates/contracts/core-types" }
 bitfun-acp = { path = "../../crates/interfaces/acp", default-features = false, features = ["client", "server"] }
 bitfun-agent-runtime = { path = "../../crates/execution/agent-runtime" }
 bitfun-agent-runtime-ipc = { path = "../../crates/adapters/agent-runtime-ipc" }
-bitfun-runtime-ports = { path = "../../crates/contracts/runtime-ports" }
+bitfun-runtime-ports = { path = "../../crates/contracts/runtime-ports", default-features = false, features = ["agent-api", "git-port", "permission", "plugin-runtime", "workspace-ports"] }
 bitfun-runtime-services = { path = "../../crates/execution/runtime-services" }
 bitfun-services-core = { path = "../../crates/services/services-core", default-features = false, features = ["dispatch-workspace", "local-storage", "process-runtime", "runtime-ownership"] }
-bitfun-agent-tools = { path = "../../crates/execution/tool-contracts" }
+bitfun-agent-tools = { path = "../../crates/execution/tool-contracts", default-features = false }
 bitfun-product-domains = { path = "../../crates/contracts/product-domains", default-features = false, features = ["external-sources"] }
 bitfun-app-server = { path = "../../crates/interfaces/app-server" }
 bitfun-app-server-client = { path = "../../crates/interfaces/app-server-client" }
