+++ src/crates/adapters/opencode-adapter/Cargo.toml
@@ -14,7 +14,7 @@ crate-type = ["rlib"]
 async-trait = { workspace = true }
 bitfun-plugin-runtime-client = { path = "../../execution/plugin-runtime-client" }
 bitfun-product-domains = { path = "../../contracts/product-domains", default-features = false, features = ["external-sources", "plugin-source"] }
-bitfun-runtime-ports = { path = "../../contracts/runtime-ports" }
+bitfun-runtime-ports = { path = "../../contracts/runtime-ports", default-features = false, features = ["plugin-runtime"] }
 bitfun-services-core = { path = "../../services/services-core", default-features = false, features = ["markdown"] }
 bitfun-static-hook-support = { path = "../static-hook-support" }
 dirs = { workspace = true }
@@ -32,6 +32,7 @@ urlencoding = { workspace = true }
 url = { workspace = true }
 
 [dev-dependencies]
+bitfun-runtime-ports = { path = "../../contracts/runtime-ports", default-features = false, features = ["script-tool-runtime"] }
 bitfun-services-integrations = { path = "../../services/services-integrations", default-features = false, features = ["plugin-source", "script-tool-runtime"] }
 tokio = { workspace = true, features = ["macros", "rt"] }
 tempfile = { workspace = true }
