+++ src/crates/assembly/product-capabilities/Cargo.toml
@@ -16,13 +16,14 @@ path = "tests/product_capability_contracts.rs"
 
 [dependencies]
 bitfun-harness = { path = "../../execution/harness" }
-bitfun-runtime-ports = { path = "../../contracts/runtime-ports" }
+bitfun-runtime-ports = { path = "../../contracts/runtime-ports", default-features = false, features = ["plugin-runtime"] }
 bitfun-runtime-services = { path = "../../execution/runtime-services" }
 bitfun-tool-packs = { path = "../../execution/tool-provider-groups", default-features = false }
 
 [dev-dependencies]
 async-trait = { workspace = true }
 bitfun-agent-runtime = { path = "../../execution/agent-runtime" }
+bitfun-runtime-ports = { path = "../../contracts/runtime-ports", default-features = false, features = ["agent-api"] }
 bitfun-runtime-services = { path = "../../execution/runtime-services", features = ["test-support"] }
 tokio = { workspace = true, features = ["macros", "rt"] }
