+++ src/crates/contracts/runtime-ports/src/runtime_event_port.rs
@@ -1,0 +1,53 @@
+use super::*;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum RuntimeEventType {
+    TurnStarted,
+    TurnCompleted,
+    TurnFailed,
+    TurnCancelled,
+    SessionStateChanged,
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RuntimeEventEnvelope {
+    pub session_id: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub turn_id: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub source: Option<AgentSubmissionSource>,
+    pub event_type: RuntimeEventType,
+    #[serde(default)]
+    pub payload: serde_json::Value,
+}
+
+#[async_trait::async_trait]
+pub trait RuntimeEventSink: Send + Sync {
+    async fn publish_runtime_event(&self, event: RuntimeEventEnvelope) -> PortResult<()>;
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn runtime_event_envelope_serializes_observational_surface_facts() {
+        let event = RuntimeEventEnvelope {
+            session_id: "session_1".to_string(),
+            turn_id: Some("turn_1".to_string()),
+            source: Some(AgentSubmissionSource::RemoteRelay),
+            event_type: RuntimeEventType::TurnCancelled,
+            payload: serde_json::json!({ "reason": "user_cancelled" }),
+        };
+
+        let json = serde_json::to_value(event).expect("serialize event");
+
+        assert_eq!(json["sessionId"], "session_1");
+        assert_eq!(json["turnId"], "turn_1");
+        assert_eq!(json["source"], "remote_relay");
+        assert_eq!(json["eventType"], "turn_cancelled");
+        assert_eq!(json["payload"]["reason"], "user_cancelled");
+    }
+}
