+++ src/crates/contracts/runtime-ports/tests/git_port_contracts.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "git-port")]
+
 use bitfun_runtime_ports::{
     WorkspaceDiffContent, WorkspaceDiffFile, WorkspaceDiffFileStatus, WorkspaceDiffSnapshot,
 };
