+++ src/crates/contracts/runtime-ports/tests/runtime_port_contracts.rs
@@ -1,10 +1,6 @@
-#[path = "runtime_port_contracts/git_port_contracts.rs"]
-mod git_port_contracts;
+#![cfg(feature = "plugin-runtime")]
+
 #[path = "runtime_port_contracts/plugin_runtime_contracts.rs"]
 mod plugin_runtime_contracts;
 #[path = "runtime_port_contracts/plugin_runtime_diagnostics_contracts.rs"]
 mod plugin_runtime_diagnostics_contracts;
-#[path = "runtime_port_contracts/script_tool_port_contracts.rs"]
-mod script_tool_port_contracts;
-#[path = "runtime_port_contracts/session_store_contracts.rs"]
-mod session_store_contracts;
