+++ src/crates/contracts/runtime-ports/tests/script_tool_port_contracts.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "script-tool-runtime")]
+
 use async_trait::async_trait;
 use bitfun_runtime_ports::{
     PortResult, ScriptToolDescriptor, ScriptToolExpectedExport, ScriptToolInvokeRequest,
