+++ src/crates/contracts/runtime-ports/tests/session_store_contracts.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "workspace-ports")]
+
 use std::path::PathBuf;
 
 use bitfun_runtime_ports::{
