+++ src/crates/interfaces/sdk-host/Cargo.toml
@@ -13,7 +13,7 @@ async-trait = { workspace = true }
 bitfun-agent-runtime = { path = "../../execution/agent-runtime" }
 bitfun-core-types = { path = "../../contracts/core-types" }
 bitfun-events = { path = "../../contracts/events" }
-bitfun-runtime-ports = { path = "../../contracts/runtime-ports" }
+bitfun-runtime-ports = { path = "../../contracts/runtime-ports", default-features = false, features = ["agent-api"] }
 futures-util = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
@@ -21,6 +21,9 @@ tokio = { workspace = true, features = ["macros", "rt", "sync", "time"] }
 tokio-util = { workspace = true, features = ["codec"] }
 tracing = { workspace = true }
 uuid = { workspace = true }
+
+[dev-dependencies]
+bitfun-runtime-ports = { path = "../../contracts/runtime-ports", default-features = false, features = ["permission"] }
 
 [lints]
 workspace = true
