+++ src/crates/services/terminal/Cargo.toml
@@ -10,7 +10,7 @@ name = "terminal_core"
 path = "src/lib.rs"
 
 [dependencies]
-bitfun-runtime-ports = { path = "../../contracts/runtime-ports" }
+bitfun-runtime-ports = { path = "../../contracts/runtime-ports", default-features = false, features = ["terminal-port"] }
 
 # Async runtime
 tokio = { workspace = true, features = ["io-util", "macros", "process", "rt", "sync", "time"] }
